android: Restructure project
All APIs go to the api package
All composables go to the composables package
